# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface UploadAudioResponse {
  sessionId: number;
  userId: string;
  audioPath: string;
  status: string;
  createdAt: string;
}

export interface SpeechScores {
  /** Words per minute score (0-100) */
  fluencyScore: number;
  /** Score based on pause frequency and duration (0-100) */
  pauseScore: number;
  /** Unique words ratio score (0-100) */
  vocabularyScore: number;
  /** Overall speaking score (0-100) */
  overallScore: number;
  wordsPerMinute: number;
  uniqueWordRatio: number;
  pauseCount: number;
  totalWords: number;
  durationSeconds: number;
}

export interface SessionAnalysis {
  sessionId: number;
  transcript: string;
  scores: SpeechScores;
  feedback: string;
  status: string;
  analyzedAt: string;
}

export interface Session {
  id: number;
  userId: string;
  audioPath: string;
  status: string;
  /** @nullable */
  transcript?: string | null;
  /** @nullable */
  fluencyScore?: number | null;
  /** @nullable */
  pauseScore?: number | null;
  /** @nullable */
  vocabularyScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  wordsPerMinute?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  createdAt: string;
  /** @nullable */
  analyzedAt?: string | null;
}

export interface ListSessionsResponse {
  sessions: Session[];
  total: number;
}

/**
 * @nullable
 */
export type SessionDetailScores = {
  /** @nullable */
  fluencyScore?: number | null;
  /** @nullable */
  pauseScore?: number | null;
  /** @nullable */
  vocabularyScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  wordsPerMinute?: number | null;
  /** @nullable */
  uniqueWordRatio?: number | null;
  /** @nullable */
  pauseCount?: number | null;
  /** @nullable */
  totalWords?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
} | null;

export interface SessionDetail {
  session: Session;
  /** @nullable */
  feedback?: string | null;
  /** @nullable */
  scores?: SessionDetailScores;
}

export interface FeedbackDetail {
  sessionId: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
  scores: SpeechScores;
  transcript: string;
  createdAt: string;
}

export interface ProgressDataPoint {
  date: string;
  sessionId: number;
  /** @nullable */
  fluencyScore?: number | null;
  /** @nullable */
  pauseScore?: number | null;
  /** @nullable */
  vocabularyScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  wordsPerMinute?: number | null;
}

export interface UserProgress {
  userId: string;
  dataPoints: ProgressDataPoint[];
  totalSessions: number;
  daysTracked: number;
}

export interface UserProgressSummary {
  userId: string;
  totalSessions: number;
  /** @nullable */
  avgFluencyScore?: number | null;
  /** @nullable */
  avgPauseScore?: number | null;
  /** @nullable */
  avgVocabularyScore?: number | null;
  /** @nullable */
  avgOverallScore?: number | null;
  /** @nullable */
  bestSession?: number | null;
  /** @nullable */
  latestSession?: number | null;
  /** Number of consecutive days with a session */
  streak: number;
  /**
   * Overall improvement percentage compared to first 3 sessions
   * @nullable
   */
  improvementPercent?: number | null;
}

export type UploadAudioBody = {
  audio: Blob;
  userId: string;
};

export type ListSessionsParams = {
  userId: string;
  limit?: number;
  offset?: number;
};

export type GetUserProgressParams = {
  days?: number;
};