# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Upload audio file for a speaking session
 */
export const UploadAudioBody = zod.object({
  audio: zod.instanceof(File),
  userId: zod.string(),
});

/**
 * @summary Analyze a speaking session
 */
export const AnalyzeSessionParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const AnalyzeSessionResponse = zod.object({
  sessionId: zod.number(),
  transcript: zod.string(),
  scores: zod.object({
    fluencyScore: zod.number().describe("Words per minute score (0-100)"),
    pauseScore: zod
      .number()
      .describe("Score based on pause frequency and duration (0-100)"),
    vocabularyScore: zod.number().describe("Unique words ratio score (0-100)"),
    overallScore: zod.number().describe("Overall speaking score (0-100)"),
    wordsPerMinute: zod.number(),
    uniqueWordRatio: zod.number(),
    pauseCount: zod.number(),
    totalWords: zod.number(),
    durationSeconds: zod.number(),
  }),
  feedback: zod.string(),
  status: zod.string(),
  analyzedAt: zod.coerce.date(),
});

/**
 * @summary List all sessions for a user
 */
export const listSessionsQueryLimitDefault = 10;
export const listSessionsQueryOffsetDefault = 0;

export const ListSessionsQueryParams = zod.object({
  userId: zod.coerce.string(),
  limit: zod.coerce.number().default(listSessionsQueryLimitDefault),
  offset: zod.coerce.number().default(listSessionsQueryOffsetDefault),
});

export const ListSessionsResponse = zod.object({
  sessions: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      audioPath: zod.string(),
      status: zod.string(),
      transcript: zod.string().nullish(),
      fluencyScore: zod.number().nullish(),
      pauseScore: zod.number().nullish(),
      vocabularyScore: zod.number().nullish(),
      overallScore: zod.number().nullish(),
      wordsPerMinute: zod.number().nullish(),
      durationSeconds: zod.number().nullish(),
      createdAt: zod.coerce.date(),
      analyzedAt: zod.coerce.date().nullish(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get a specific session with transcript, scores, and feedback
 */
export const GetSessionParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  session: zod.object({
    id: zod.number(),
    userId: zod.string(),
    audioPath: zod.string(),
    status: zod.string(),
    transcript: zod.string().nullish(),
    fluencyScore: zod.number().nullish(),
    pauseScore: zod.number().nullish(),
    vocabularyScore: zod.number().nullish(),
    overallScore: zod.number().nullish(),
    wordsPerMinute: zod.number().nullish(),
    durationSeconds: zod.number().nullish(),
    createdAt: zod.coerce.date(),
    analyzedAt: zod.coerce.date().nullish(),
  }),
  feedback: zod.string().nullish(),
  scores: zod
    .object({
      fluencyScore: zod.number().nullish(),
      pauseScore: zod.number().nullish(),
      vocabularyScore: zod.number().nullish(),
      overallScore: zod.number().nullish(),
      wordsPerMinute: zod.number().nullish(),
      uniqueWordRatio: zod.number().nullish(),
      pauseCount: zod.number().nullish(),
      totalWords: zod.number().nullish(),
      durationSeconds: zod.number().nullish(),
    })
    .nullish(),
});

/**
 * @summary Get AI feedback for a specific session
 */
export const GetSessionFeedbackParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const GetSessionFeedbackResponse = zod.object({
  sessionId: zod.number(),
  feedback: zod.string(),
  strengths: zod.array(zod.string()),
  improvements: zod.array(zod.string()),
  scores: zod.object({
    fluencyScore: zod.number().describe("Words per minute score (0-100)"),
    pauseScore: zod
      .number()
      .describe("Score based on pause frequency and duration (0-100)"),
    vocabularyScore: zod.number().describe("Unique words ratio score (0-100)"),
    overallScore: zod.number().describe("Overall speaking score (0-100)"),
    wordsPerMinute: zod.number(),
    uniqueWordRatio: zod.number(),
    pauseCount: zod.number(),
    totalWords: zod.number(),
    durationSeconds: zod.number(),
  }),
  transcript: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get user progress and speaking improvement over time
 */
export const GetUserProgressParams = zod.object({
  userId: zod.coerce.string(),
});

export const getUserProgressQueryDaysDefault = 30;

export const GetUserProgressQueryParams = zod.object({
  days: zod.coerce.number().default(getUserProgressQueryDaysDefault),
});

export const GetUserProgressResponse = zod.object({
  userId: zod.string(),
  dataPoints: zod.array(
    zod.object({
      date: zod.coerce.date(),
      sessionId: zod.number(),
      fluencyScore: zod.number().nullish(),
      pauseScore: zod.number().nullish(),
      vocabularyScore: zod.number().nullish(),
      overallScore: zod.number().nullish(),
      wordsPerMinute: zod.number().nullish(),
    }),
  ),
  totalSessions: zod.number(),
  daysTracked: zod.number(),
});

/**
 * @summary Get summarized stats and improvement highlights for a user
 */
export const GetUserProgressSummaryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserProgressSummaryResponse = zod.object({
  userId: zod.string(),
  totalSessions: zod.number(),
  avgFluencyScore: zod.number().nullish(),
  avgPauseScore: zod.number().nullish(),
  avgVocabularyScore: zod.number().nullish(),
  avgOverallScore: zod.number().nullish(),
  bestSession: zod.number().nullish(),
  latestSession: zod.number().nullish(),
  streak: zod.number().describe("Number of consecutive days with a session"),
  improvementPercent: zod
    .number()
    .nullish()
    .describe("Overall improvement percentage compared to first 3 sessions"),
});