# lib/api-zod/src/generated/types/feedbackDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeechScores } from "./speechScores";

export interface FeedbackDetail {
  sessionId: number;
  feedback: string;
  strengths: string[];
  improvements: string[];
  scores: SpeechScores;
  transcript: string;
  createdAt: Date;
}