# lib/api-zod/src/generated/types/listSessionsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { Session } from "./session";

export interface ListSessionsResponse {
  sessions: Session[];
  total: number;
}