# lib/api-zod/src/generated/types/progressDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-Based Daily Speaking Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressDataPoint {
  date: Date;
  sessionId: number;
  /** @nullable */
  fluencyScore?: number | null;
  /** @nullable */
  pauseScore?: number | null;
  /** @nullable */
  vocabularyScore?: number | null;
  /** @nullable */
  overallScore?: number | null;
  /** @nullable */
  wordsPerMinute?: number | null;
}